% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctCentrality.R
\name{ctCentrality}
\alias{ctCentrality}
\title{Calculate CT Centrality measures}
\usage{
ctCentrality(drift, dt)
}
\arguments{
\item{drift}{input \eqn{p \times p} drift matrix}

\item{dt}{the time-interval \eqn{\Delta t}}
}
\value{
a list containing the centrality measures for each variable
}
\description{
Calculate CT Centrality measures
}
\examples{
drift <- matrix(c(-.357, 0, 0,
                 .771, -.511, 0,
                 -.450, .729, -.693), 3, 3, byrow = TRUE)
dt = 2
ct_centrality(drift=drift, dt=dt)
}
\seealso{
\code{\link{DE}}, \code{\link{TE}}, \code{\link{IE}}
}
