% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctCentrality_int.R
\name{ctCentrality_int}
\alias{ctCentrality_int}
\title{Calculate CT Centrality measures (internal function)}
\usage{
ctCentrality_int(drift, dt, listout = TRUE)
}
\arguments{
\item{drift}{input \eqn{p \times p} drift matrix}

\item{dt}{the time-interval \eqn{\Delta t}}

\item{listout}{if TRUE output is a list, else a vector}
}
\value{
a list or vector containing the centrality measures for each variable.
}
\description{
Calculate CT Centrality measures (internal function)
}
\seealso{
\code{\link{DE}}, \code{\link{TE}}, \code{\link{IE}}
}
\keyword{Internal}
