% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE.R
\name{DE}
\alias{DE}
\title{Calculate CT Direct Effect \eqn{DE(\Delta t)}}
\usage{
DE(drift, dt, IV, DV, M)
}
\arguments{
\item{drift}{input \eqn{p \times p} drift matrix}

\item{dt}{the time-interval \eqn{\Delta t}}

\item{IV}{index corresponding to the independent-variable}

\item{DV}{index corresponding to the dependent-variable}

\item{M}{index or indices corresponding to the mediators to be intervened on}
}
\value{
a scalar direct effect corresponding to a continuous intervention on M
}
\description{
Calculate CT Direct Effect \eqn{DE(\Delta t)}
}
\examples{
drift<- matrix(c(-.357, 0, 0,
                 .771, -.511, 0,
                 -.450, .729, -.693), 3, 3, byrow = TRUE)
dt = 2
IV = 1
DV = 3
M = 2
DE(drift=drift,dt=dt,IV=IV,DV=DV,M=M)
}
\seealso{
\code{\link{DE}}
}
