% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCIs.R
\name{getCIs}
\alias{getCIs}
\title{Calculate confidence/credible intervals of arbitrary functions of the drift matrix}
\usage{
getCIs(
  posterior,
  simplify = TRUE,
  probs = c(0.025, 0.5, 0.975),
  const = 1,
  FUN,
  ...
)
}
\arguments{
\item{posterior}{array of dimension \eqn{s \times p \times p} giving s posterior samples
of a \eqn{p \times p} drift matrix}

\item{simplify}{argument to be passed to sapply. Defaults to TRUE}

\item{probs}{vector of quantiles to be computed. By default returns the 2.5, 50 and 97.5 percentile}

\item{const}{any number with which to multiply the drift matrix. Default to 1. Should only be used
in combination with the expm function, to allow for computation of \eqn{\Phi(\Delta t = const)}}

\item{FUN}{any function that takes a drift matrix as input and returns a vector}

\item{...}{arguments to be passed to FUN}
}
\value{
a matrix or vector containing lower bound, point estimate and upper bound quantiles
}
\description{
Calculate confidence/credible intervals of arbitrary functions of the drift matrix
}
\seealso{
\code{\link{DE}}, \code{\link{TE}}, \code{\link{IE}}
}
