% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IE.R
\name{IE}
\alias{IE}
\title{Calculate CT Indirect Effect \eqn{IE(\Delta t)}}
\usage{
IE(drift, dt, IV, DV, M)
}
\arguments{
\item{drift}{input \eqn{p \times p} drift matrix}

\item{dt}{the time-interval \eqn{\Delta t}}

\item{IV}{index corresponding to the independent-variable}

\item{DV}{index corresponding to the dependent-variable}

\item{M}{index or indices corresponding to the mediators}
}
\value{
a scalar direct effect corresponding to a continuous intervention on M
}
\description{
Calculate CT Indirect Effect \eqn{IE(\Delta t)}
}
\examples{
drift<- matrix(c(-.357, 0, 0,
                 .771, -.511, 0,
                 -.450, .729, -.693), 3, 3, byrow = TRUE)
dt = 2
IV = 1
DV = 3
M = 2
occ <- 1
IE(drift=drift,dt=dt,IV=IV,DV=DV,M=M)
}
\seealso{
\code{\link{DE}}, \code{\link{TE}}
}
