% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCentrality.R
\name{plotCentrality}
\alias{plotCentrality}
\title{Function to plot centrality metrics across a range of time-intervals}
\usage{
plotCentrality(
  drift = NULL,
  posterior = NULL,
  CI_obj = NULL,
  dts = seq(0, 1, 0.5),
  plot = TRUE,
  type = "TEC",
  x = 1
)
}
\arguments{
\item{drift}{a \eqn{p \times p} drift matrix. Can be a matrix of point estimates. Optional input, not necessary if 
`posterior` or `CI_obj` supplied. Defaults to NULL}

\item{posterior}{array of dimension \eqn{s \times p \times p} giving s posterior samples
of a \eqn{p \times p} drift matrix. Must be supplied if `CI_obj` absent. Default is NULL}

\item{CI_obj}{a \eqn{3 \times 8 \times t} array containing lower bound, upper bound, and point estimates of each centrality metric
across a range of \eqn{t} time-intervals, as output by the `getCIs()` function or an earlier function call. See `plot`and details.}

\item{dts}{a vector of \eqn{t} time-intervals. If `CI_obj` is supplied, must have some dimensions as used in generation}

\item{plot}{logical return a plot or array object of type `CI_obj`. See details}

\item{type}{character string. "TEC" plots the total effect centrality, "IEC" the indirect effect centrality and "both" both.}

\item{x}{scalar or vector giving the index of the variable whose centrality you wish to plot. Defaults to 1}
}
\value{
Returns a plot by default. If  `plot` is set to FALSE and `posterior` is supplied, returns an object of type `CI_obj`. This
can be used to do all necessary (slow) operations first, then used as input for a future plotting call calculations
}
\description{
Function to plot centrality metrics across a range of time-intervals
}
\seealso{
\code{\link{ctCentrality}}, \code{\link{plotPhi}}
}
