% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPhi.R
\name{plotPhi}
\alias{plotPhi}
\title{Visualize the model-implied lagged effects across a range of time-intervals}
\usage{
plotPhi(
  drift = NULL,
  posterior = NULL,
  CI_obj = NULL,
  dts = seq(0, 1, 0.5),
  plot = TRUE,
  index = "all",
  colvec = NULL,
  leg = TRUE,
  poly = TRUE,
  drawaxis = TRUE,
  main = "Lagged effects as a function of Time-Interval",
  xlab = "Time-Interval",
  ylab = "Effect Size",
  ind = NULL,
  cex.lab = 1,
  legtext = NULL
)
}
\arguments{
\item{drift}{a \eqn{p \times p} drift matrix. Can be a matrix of point estimates. Optional input, not necessary if 
`posterior` or `CI_obj` supplied. Defaults to NULL}

\item{posterior}{array of dimension \eqn{s \times p \times p} giving s posterior samples
of a \eqn{p \times p} drift matrix. Must be supplied if `CI_obj` absent. Default is NULL}

\item{CI_obj}{a \eqn{3 \times p^2 \times t} array containing lower bound, upper bound, and point estimates of each lagged parameter
across a range of \eqn{t} time-intervals, as output by the `getCIs()` function or an earlier function call. See `plot`and details.}

\item{dts}{a vector of \eqn{t} time-intervals. If `CI_obj` is supplied, must have some dimensions as used in generation}

\item{plot}{logical return a plot or array object of type `CI_obj`. See details}

\item{index}{matrix or character vector describing which parameters to plot. The matrix must be of dimension \eqn{q \times 2}
giving the row and column of the `q` parameters to be plotted. Character strings accepted are: \cr
"all": plot all lagged parameters \cr
"AR": plot only auto-regressive (diagonal) parameters \cr
"CL": plot only cross-lagged (off-diagonal) parameters}

\item{colvec}{vector of length `q` of colors to be used for plotting. Defaults to NULL. If not supplied, function will plot
a spectrum of `q` colors from blue to red using `colorRampPallete()`}

\item{leg}{logical. Include a legend or not. Defaults to TRUE}

\item{poly}{logical. Draw a shaded polygon to represent the CIs}

\item{cex.lab}{text size for labels}

\item{legtext}{Default option NULL generates custom legend text using greek Phi notation and subscripts. Otherwise, supply text which be parsed by as.expression, e.g. a list of bquote text}

\item{makeaxis}{logical. Draw x and y axis and ticks. Defaults to TRUE}

\item{main.}{Title for the plot}

\item{xlab.}{x-axis label}

\item{ylab.}{y-axis label}

\item{ind.}{Optional argument, which indices to plot.}
}
\value{
Returns a plot by default. If  `plot` is set to FALSE and `posterior` is supplied, returns an object of type `CI_obj`. This
can be used to do all necessary (slow) operations first, then used as input for a future plotting call calculations
}
\description{
Visualize the model-implied lagged effects across a range of time-intervals
}
\seealso{
\code{\link{plotCentrality}}
}
