% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\docType{data}
\name{simdata}
\alias{simdata}
\title{Simulated data from a single-subject four-variable CT-VAR(1) model}
\format{
A data frame with 9999 rows and 6 variables:
\describe{
  \item{id}{id of participant. All rows are 1}
  \item{time}{time since the start of the study. All time-intervals between subsequent observations were generated
  using a uniform disribution bounded by .001 and 2}
  \item{Y1}{First of the four time-series variables}
  \item{Y2}{Second of the four time-series variables}
  \item{Y3}{Third of the four time-series variables}
  \item{Y4}{Fourth of the four time-series variables}
  ...
}
}
\usage{
simdata
}
\description{
Simulated data from a single-subject four-variable CT-VAR(1) model
}
\keyword{datasets}
