% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPress.R
\name{simPress}
\alias{simPress}
\title{Calculate the effect of a press intervention}
\usage{
simPress(drift, dt, M, press, start = NULL)
}
\arguments{
\item{drift}{input \eqn{p \times p} drift matrix}

\item{dt}{the time-interval considered \eqn{\Delta t}}

\item{M}{index or indices corresponding to the intervened-on variables}

\item{press}{scalar or vector giving the value of the press intervention(s)}

\item{start}{(optional) full vector of starting values.}
}
\value{
a vector representing the value of the CT process after a press intervention to set M to value press
}
\description{
Calculate the effect of a press intervention
}
\examples{
drift <- matrix(c(-1.2,  .25,     0,  1,
                 1.1, -.5,     0,   0,
                   0,  1.18, -1.2,   0 ,
                 0, -1.46,     .5, -1.2),4,4,byrow=T)

M = 2; press = 1; start = c(0,M,0,0)
 # Set up the time vector
   dts2 <- seq(0,5,.001)
 # Calculate effect over a range of time intervals
   y_ct <- matrix(0,length(dts2),n+1)

# Get Trajectories for CT and total (DT2) effects

 for(i in 1:length(dts2)){
 # Total Effect
 y_ct[i,1:n] <- simPress(drift,dt=dts2[i],M,press, start)
 y_ct[i,n+1] <- dts2[i]
}

# Create plot
cols <- c("#1B9E77", "#D95F02", "#7570B3", "#E7298A")
plot.new()
plot.window()plot.window(xlim=c(0, 5), ylim=c(-1,1))
for(c in 1:4) lines(y = y_ct[,c], x = y_ct[,5], type = "l", col = cols[c], lwd = lwd)
abline(h = 0); axis(1); axis(2); title(xlab = "Time (t)", ylab = "Variable Values")

# To get the new equilibrium plug in a large value for Delta t
simPress(drift,dt=100,M,press, start)

}
\seealso{
\code{\link{DE}}
}
