% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TE.R
\name{TE}
\alias{TE}
\title{CT Total Effect}
\usage{
TE(drift, dt, IV, DV)
}
\arguments{
\item{drift}{input \eqn{p \times p} drift matrix}

\item{dt}{the time-interval \eqn{\Delta t}}

\item{IV}{index corresponding to the independent-variable}

\item{DV}{index corresponding to the dependent-variable}
}
\value{
a scalar total effect
}
\description{
CT Total Effect
}
\examples{
drift<- matrix(c(-.357, 0, 0,
                 .771, -.511, 0,
                 -.450, .729, -.693), 3, 3, byrow = TRUE)
dt = 2
IV = 1
DV = 3
TE(drift=drift,dt=dt,IV=IV,DV=DV)
}
