% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctCentrality.R
\name{ctCentrality}
\alias{ctCentrality}
\title{Calculate CT Centrality measures with CIs}
\usage{
ctCentrality(
  drift = NULL,
  dt,
  listout = FALSE,
  posterior = NULL,
  probs = c(0.025, 0.5, 0.975)
)
}
\arguments{
\item{drift}{input \eqn{p \times p} drift matrix}

\item{dt}{the time-interval \eqn{\Delta t}}

\item{listout}{if TRUE output is a list, else a vector}

\item{posterior}{optional. Default is NULL. Users can supply an array of dimension
\eqn{s \times p \times p} of samples from the posterior of the drift matrix, 
as output from `ctsem::ctExtract()$pop_DRIFT`. If this is supplied
then the function returns quantiles of centrality metrics from the posterior, as defined by `probs`}

\item{probs}{vector of probabilities to be passed to `quantile()`. By default set to
return mean (50th percentile) and 95 percent CIs (2.5 and 97.5 percentiles)}
}
\value{
a list or vector containing the centrality measures for each variable. If posterior is supplied,
then returns an array of \eqn{1 \times 2p \times 3}, with slices point estimate, lower and upper bound
}
\description{
Calculate CT Centrality measures with CIs
}
\examples{
drift <- matrix(c(-.357, 0, 0,
                 .771, -.511, 0,
                 -.450, .729, -.693), 3, 3, byrow = TRUE)
dt = 2
ctCentrality(drift = drift, dt = dt)
}
\seealso{
\code{\link{DE}}, \code{\link{TE}}, \code{\link{IE}}, \code{\link{getCIs}}
}
